requirements.txt
pyspark==4.0.1
kafka-python